Fix FastAPI dependency loading.
Newer versions of FastAPI have managed to break the use of
`StreamingResponse` with `yield` dependencies. To resolve
this, I have refactored the code to not use `yield`
dependencies at all. Basically, it just keeps a single
connection to each backend open in each process and reuses
it over and over again. This might have the side-effect of
improving performance.